[MC-4] [api] refactor: change file name
