+++ apps/api/src/numerical-guidance/api/major-chart/dto/get-major-chart.dto.ts
@@ -1,5 +1,5 @@
 import { ApiProperty } from '@nestjs/swagger';
-import ChartTimeline from './char-timeline.dto';
+import ChartTimeline from './chart-timeline.dto';
 
 export class GetMajorChartDto {
   @ApiProperty({
